適当

+++ vim_link.sh
@@ -6,20 +6,26 @@ TARGET_DIR="$CURRENT_DIR/.config/nvim"
 NVIM_DIR="$HOME/.config/nvim"
 
 if [ -d "$NVIM_DIR" ]; then
-    BACKUP_DIR="$HOME/.config/nvim_backup_$(date +"%Y%m%d_%H%M%S")"
-    mv "$NVIM_DIR" "$BACKUP_DIR"
-    echo "backup done"
+    BACKUP_DIR="$HOME/.config/nvim/nvim_backup_$(date +"%Y%m%d_%H%M%S")"
+    rsync -a "$NVIM_DIR" "$BACKUP_DIR"
+    echo "nvim backup done"
+else
+	mkdir -p "$NVIM_DIR"
+	for item in "$TARGET_DIR"/*; do
+		if [ -e "$item" ]; then
+			ln -sfn "$item" "$NVIM_DIR/$(basename "$item")"
+		fi
+	done
+	echo "nvim linked"
 fi
 
-mkdir -p "$NVIM_DIR"
-
-for item in "$TARGET_DIR"/*; do
-    if [ -e "$item" ]; then
-        ln -sfn "$item" "$NVIM_DIR/$(basename "$item")"
-    fi
-done
-
 # vim link
-ln -sfn "$CURRENT_DIR/.vimrc" "$HOME/.vimrc"
+if [ -f "$CURRENT_DIR/.vimrc" ]; then
+    BACKUP_FILE="$HOME/.vimrc_backup_$(date +"%Y%m%d_%H%M%S")"
+	rsync -oL "$HOME/.vimrc" "$BACKUP_FILE"
+	echo ".vimrc backup done"
+else
+	ln -sfn "$CURRENT_DIR/.vimrc" "$HOME/.vimrc"
+	echo ".vimrc linked"
+fi
 
-echo "linked"
